Implemented buddyMemoryManager

+++ Kernel/memoryManager/listFreeMemoryManager.c
@@ -45,7 +45,7 @@ static BlockLink_t xStart, xEnd;
 		pxIterator->pxNextFreeBlock = pxBlockToInsert;                                                                             \
 	}
 
-void *malloc(unsigned int xWantedSize)
+void *malloc(size_t xWantedSize)
 {
 	BlockLink_t *pxBlock, *pxPreviousBlock, *pxNewBlockLink;
 	static int xHeapHasBeenInitialised = 0;
